+++ src/array/array_sync.rs
@@ -190,15 +190,16 @@ impl<TStorage: ?Sized + ReadableStorageTraits> Array<TStorage> {
         };
 
         // Decode the subset of the chunk which intersects array_subset
+        let overlap = unsafe { array_subset.overlap_unchecked(&chunk_subset_in_array) };
         let array_subset_in_chunk_subset =
-            unsafe { array_subset.in_subset_unchecked(&chunk_subset_in_array) };
+            unsafe { overlap.relative_to_unchecked(chunk_subset_in_array.start()) };
         let decoded_bytes =
             self.retrieve_chunk_subset(chunk_indices, &array_subset_in_chunk_subset)?;
 
         // Copy decoded bytes to the output
         let element_size = self.data_type().size() as u64;
         let chunk_subset_in_array_subset =
-            unsafe { chunk_subset_in_array.in_subset_unchecked(array_subset) };
+            unsafe { overlap.relative_to_unchecked(array_subset.start()) };
         let mut decoded_offset = 0;
         for (array_subset_element_index, num_elements) in unsafe {
             chunk_subset_in_array_subset
@@ -820,8 +821,9 @@ impl<TStorage: ?Sized + ReadableStorageTraits + WritableStorageTraits> Array<TSt
                 // This skips the internal decoding occurring in store_chunk_subset
                 self.store_chunk(chunk_indices, subset_bytes)?;
             } else {
+                let overlap = unsafe { array_subset.overlap_unchecked(&chunk_subset_in_array) };
                 let chunk_subset_in_array_subset =
-                    unsafe { chunk_subset_in_array.in_subset_unchecked(array_subset) };
+                    unsafe { overlap.relative_to_unchecked(array_subset.start()) };
                 let chunk_subset_bytes = unsafe {
                     chunk_subset_in_array_subset.extract_bytes_unchecked(
                         &subset_bytes,
@@ -832,7 +834,7 @@ impl<TStorage: ?Sized + ReadableStorageTraits + WritableStorageTraits> Array<TSt
 
                 // Store the chunk subset
                 let array_subset_in_chunk_subset =
-                    unsafe { array_subset.in_subset_unchecked(&chunk_subset_in_array) };
+                    unsafe { overlap.relative_to_unchecked(chunk_subset_in_array.start()) };
 
                 self.store_chunk_subset(
                     chunk_indices,
@@ -847,8 +849,9 @@ impl<TStorage: ?Sized + ReadableStorageTraits + WritableStorageTraits> Array<TSt
                         .subset_unchecked(&chunk_indices, self.shape())
                         .unwrap()
                 };
+                let overlap = unsafe { array_subset.overlap_unchecked(&chunk_subset_in_array) };
                 let chunk_subset_in_array_subset =
-                    unsafe { chunk_subset_in_array.in_subset_unchecked(array_subset) };
+                    unsafe { overlap.relative_to_unchecked(array_subset.start()) };
                 let chunk_subset_bytes = unsafe {
                     chunk_subset_in_array_subset.extract_bytes_unchecked(
                         &subset_bytes,
@@ -859,7 +862,7 @@ impl<TStorage: ?Sized + ReadableStorageTraits + WritableStorageTraits> Array<TSt
 
                 // Store the chunk subset
                 let array_subset_in_chunk_subset =
-                    unsafe { array_subset.in_subset_unchecked(&chunk_subset_in_array) };
+                    unsafe { overlap.relative_to_unchecked(chunk_subset_in_array.start()) };
 
                 self.store_chunk_subset(
                     &chunk_indices,